Repair Ship Nanotool working with POBs
Use base.cfg values for repairship_chance and repairship_rate to control.

This consumes repair commodities from the repairship and will not heal a POB without them. Nanotool will also not trigger a POB's shield.

PlayerRestrictions.cpp in Alley also updated as I could only get the Nanotool to work on POBs with a damage value of 2. This can be reverted if there is something that I overlooked that prevented the Nanotool from registering on POBs when it only had 1 damage.