Adds Shield Piercing capability
This is done through a piercing multiplier that is applied to the damage that was done to the shield. Below example allows 75% of the Corona's hull damage to bypass shields while they're up. BCr Breaker in the Config Example does not penetrate shields.

Config Example

[EquipAndPiercingWeapons]
; munition = equipment damage multiplier, piercing damage multiplier, shield piercing
dsy_codebc_breaker_ammo = 10.0, 2.0, 0.0
torpedo01_mark022_ammo = 1.0, 1.0, 1.5