Initial commit for CombatLog plugin
This includes a change to FLHook\HkCbDeath.cpp and FLHookPluginSDK\headers\plugin.h to include Plugin_Communication with the CombatLog to be able to override self-kills.

The CombatLog by itself tracks player damage dealt / received and displays this upon docking or being killed.

Self-Kill Override works as follows:
If the ClientID of the victim and the killer are the same, FLHook makes a plugin communication call to Combat Log.
This searches the dead player's damage received log and returns the ClientID of the player that had the highest damage dealt to the dead player and is within 15k range.

iClientIDKiller is then updated to the returned ClientID with the cause indicating a "Self-Kill Override".